Crate asset_classification_smart_contract

Asset Classification Smart Contract

This contract uses Cosmwasm’s provided architecture in conjunction with Provwasm to create a wasm smart contract that can be deployed to and interact with the Provenance Blockchain.

Modules

contract

The entrypoint for all external commands sent to the compiled wasm.

core

Contains all types and base functionality used to construct the logic of the contract.

execute

Contains all execution routes used by the contract file.

instantiate

Contains the functionality used in the contract file to instantiate the contract.

migrate

Contains the functionality used in the contract file to perform a contract migration.

query

Contains the functionality used in the contract file to perform a contract query.

service

Complex structs used to perform intensive operations in a centralized location.

util

Miscellaneous functionalities that do not logically belong to a concrete group.

validation

Functionality used to ensure the logical integrity of received external values.