Enum asset_classification_smart_contract::core::error::ContractError

pub enum ContractError {

    Bech32Error(Error),
    SemVer(Error),
    Std(StdError),
    UuidError(Error),
    AssetAlreadyOnboarded {
        scope_address: String,
    },
    AssetAlreadyVerified {
        scope_address: String,
        status: AssetOnboardingStatus,
    },
    AssetNotFound {
        scope_address: String,
    },
    AssetPendingVerification {
        scope_address: String,
        verifier_address: String,
    },
    AssetSpecMismatch {
        asset_type: String,
        scope_address: String,
        scope_spec_address: String,
        expected_scope_spec_address: String,
    },
    AssetTypeDisabled {
        asset_type: String,
    },
    DuplicateVerifierProvided,
    GenericError {
        msg: String,
    },
    InvalidAddress {
        address: String,
        explanation: String,
    },
    InvalidContractName {
        current_contract: String,
        migration_contract: String,
    },
    InvalidContractVersion {
        current_version: String,
        migration_version: String,
    },
    InvalidFunds(String),
    InvalidMessageFields {
        message_type: String,
        invalid_fields: Vec<String>,
    },
    InvalidMessageType {
        expected_message_type: String,
    },
    InvalidScope {
        explanation: String,
    },
    NotFound {
        explanation: String,
    },
    RecordAlreadyExists {
        explanation: String,
    },
    RecordNotFound {
        explanation: String,
    },
    Unauthorized {
        explanation: String,
    },
    UnauthorizedAssetVerifier {
        scope_address: String,
        verifier_address: String,
        expected_verifier_address: String,
    },
    UnexpectedSerializedEnum {
        received_type: String,
        explanation: String,
    },
    UnexpectedState {
        explanation: String,
    },
    Unimplemented,
    UnsupportedAssetType {
        asset_type: String,
    },
    UnsupportedVerifier {
        verifier_address: String,
        asset_type: String,
    },

}

A massive enum including all the different error scenarios that can be encountered throughout each process in the contract.

Variants

Bech32Error(Error)

An interceptor for a Bech32 Error.

SemVer(Error)

An interceptor for a SemVer Error.

Std(StdError)

An interceptor for a Cosmwasm Error.

UuidError(Error)

An interceptor for a Uuid Error.

AssetAlreadyOnboarded

Fields

scope_address: String

The bech32 scope address of the already-onboarded asset.

This error is encountered when an asset is attempted in the onboarding process, but it has already been onboarded and classified.

AssetAlreadyVerified

Fields

scope_address: String

The bech32 address of the scope that has already been verified.

status: AssetOnboardingStatus

The current onboarding status in the AssetScopeAttribute on the scope that has already been verified.

An error emitted when a verifier attempts to run the verification process on an asset that does not require it. This can only occur when a verifier attempts to run a duplicate verification on a scope that has already had its verification process completed.

AssetNotFound

Fields

scope_address: String

The bech32 address of the scope that does not appear to yet exist on the Provenance Blockchain.

This error is encountered when the onboarding process cannot locate the scope specified by the requestor.

AssetPendingVerification

Fields

scope_address: String

The bech32 scope address of the asset that has been onboarded.

verifier_address: String

The bech32 address of the verifier that will perform verification on the asset.

This error is encountered when the onboarding process is attempted for an asset that has already been onboarded as is awaiting a decision from its target verifier.

AssetSpecMismatch

Fields

asset_type: String

The type of asset provided by the requestor for onboarding.

scope_address: String

The bech32 scope address of the asset provided during onboarding.

scope_spec_address: String

The bech32 scope specification address listed on the scope provided during onboarding.

expected_scope_spec_address: String

The bech32 scope specification address listed in the AssetDefinition stored for the given asset type.

This error is encountered when an asset is attempted to be onboarded as a specific asset_type, but the AssetDefinition for that type has a different Provenance Blockchain Scope Specification bech32 address listed in the contract’s internal storage. This is to prevent scopes from being onboarded as a specific type unless they meet the correct specification listed on the blockchain.

AssetTypeDisabled

Fields

asset_type: String

The type of asset that is currently disabled.

This error indicates that an asset was attempted to be onboarded with an asset_type linked to an AssetDefinition that is currently not enabled]

DuplicateVerifierProvided

Denotes that an existing VerifierDetail has the same address property as the provided VerifierDetail to be added to an AssetDefinition.

GenericError

Fields

msg: String

A free-form text description of the error that occurred.

An error that can be used in a circumstance where a named error is not necessary to be created.

InvalidAddress

Fields

address: String

The invalid bech32 address.

explanation: String

A message further explaining the issue.

Indicates that a bech32 address was provided that does not meet proper specifications for the given scenario.

InvalidContractName

Fields

current_contract: String

The name of the existing contract. Should correlate to the value specified in Cargo.toml’s name property.

migration_contract: String

The name of the incorrect contract.

An error that can occur during a migration that indicates that an incorrect stored contract code was attempted to be provided for a migration.

InvalidContractVersion

Fields

current_version: String

The version of the contract currently active on the Provenance Blockchain.

migration_version: String

The version of the stored code used in the migration.

An error that can occur during a migration that indicates that the stored contract code that was used as the migration target has a version that is lower than the current contract version. This error is to prevent downgrading the contract to a previous version, which could remove features and cause problems.

InvalidFunds(String)

Tuple Fields

0: String

Denotes the reason that invalid funds were detected.

A generic error that specifies that some form of provided or utilized coin was invalid.

InvalidMessageFields

Fields

message_type: String

Indicates the type of message that was sent, causing the error. Will be one of the values in the msg file.

invalid_fields: Vec<String>

A collection of messages that indicate every issue present in the bad msg.

An error emitted by the validation that indicates that input values were not properly specified (blank strings, invalid numbers, etc).

InvalidMessageType

Fields

expected_message_type: String

Denotes the correct msg that was expected to be provided.

An error emitted when an internal issue arises, indicating that an incorrect msg was sent to a function receiver. The functions that use this error are all at first called into in the contract file.

InvalidScope

Fields

explanation: String

A free-form text description of the reason that the scope is considered invalid.

An error that indicates that a scope inspected during the onboarding process is missing internal values and is not valid for onboarding, like an internal Provenance Blockchain Metadata Record.

NotFound

Fields

explanation: String

A message describing the reason that the resource could not be found.

An error that occurs when a lookup is attempted for a contract resource but the resource does not exist. For instance, when an AssetDefinition does not contain a VerifierDetail with a specified bech32 address, this error will occur.

RecordAlreadyExists

Fields

explanation: String

A free-form text description of the reason that the record already exists.

An error that occurs when a unique key is violated during an attempt to add new data to the contract’s internal storage. Reference: state.

RecordNotFound

Fields

explanation: String

A free-form text description of the record that could not be found.

Occurs when a mandatory data lookup is performed on the contract’s internal storage, but the required value is not found. Reference: state.

Unauthorized

Fields

explanation: String

A free-form text description of why the action was not authorized.

A generic error that occurs when an address attempts to perform an operation in the contract that it does not have the permission to.

UnauthorizedAssetVerifier

Fields

scope_address: String

The bech32 address of the scope that is awaiting verification.

verifier_address: String

The bech32 address of the account that attempted to run verification.

expected_verifier_address: String

The bech32 address of the account that has been requested to run verification.

An error emitted when an account attempts to initiate the verification process for a scope when its bech32 address is not listed as the verifier for the particular onboarding instance.

UnexpectedSerializedEnum

Fields

received_type: String

The type value of the serialized enum that could not be properly converted.

explanation: String

A free-form text description of what went wrong in converting the value to its underlying type.

This error occurs when a SerializedEnum is received from a caller that cannot be properly converted to its expected underlying type.

UnexpectedState

Fields

explanation: String

A free-form text description of why the operation was rejected.

An error that can occur when the contract is in an unexpected state and refuses to perform an operation.

Unimplemented

A placeholder error that can be used as a stopgap during the implementation of functionality to allow the tests to compile and run before the feature is completed. This error should never be used in a release build.

UnsupportedAssetType

Fields

asset_type: String

The type of asset that could not be located for onboarrding.

This error is encountered when an asset type is attempted to be used during the onboarding process, but the asset type cannot be found in the contract’s internal storage.

UnsupportedVerifier

Fields

verifier_address: String

The bech32 address of the target verifier.

asset_type: String

The asset_type selected during onboarding.

This error can occur when a target VerifierDetail does not exist in an AssetDefinition during the onboarding process.

Implementations

impl ContractError

pub fn generic<S: Into<String>>(msg: S) -> ContractError

Constructs an instance of the GenericError variant, allowing a generic Into implementation.

Parameters

msg The string value to use as the msg field of the generic error.