Enum asset_classification_smart_contract::core::msg::ExecuteMsg

pub enum ExecuteMsg {
    OnboardAsset {
        identifier: SerializedEnum,
        asset_type: String,
        verifier_address: String,
        access_routes: Option<Vec<AccessRoute>>,
        add_os_gateway_permission: Option<bool>,
    },
    VerifyAsset {
        identifier: SerializedEnum,
        asset_type: String,
        success: bool,
        message: Option<String>,
        access_routes: Option<Vec<AccessRoute>>,
    },
    AddAssetDefinition {
        asset_definition: AssetDefinitionInputV3,
    },
    UpdateAssetDefinition {
        asset_definition: AssetDefinitionInputV3,
    },
    ToggleAssetDefinition {
        asset_type: String,
        expected_result: bool,
    },
    AddAssetVerifier {
        asset_type: String,
        verifier: VerifierDetailV2,
    },
    UpdateAssetVerifier {
        asset_type: String,
        verifier: VerifierDetailV2,
    },
    UpdateAccessRoutes {
        identifier: SerializedEnum,
        asset_type: String,
        owner_address: String,
        access_routes: Vec<AccessRoute>,
    },
    DeleteAssetDefinition {
        asset_type: String,
    },
}

Defines all routes in which the contract can be executed. These are all handled directly in the contract file.

Variants

OnboardAsset

This route is the primary interaction point for most consumers. It consumes an asset uuid or scope address, the type of asset corresponding to that scope (heloc, mortgage, payable, etc), and, if all checks pass, attaches an attribute to the provided scope that stages the scope for verification of authenticity by the specified verifier in the request. The attribute is attached based on the base_contract_name specified in the contract, combined with the specified asset type in the request. Ex: if base_contract_name is “asset” and the asset type is “myasset”, the attribute would be assigned under the name of “myasset.asset”. All available asset types are queryable, and stored in the contract as AssetDefinitionV3 values. After onboarding is completed, an AssetScopeAttribute will be stored on the scope with an AssetOnboardingStatus of Pending, indicating that the asset has been onboarded to the contract but is awaiting verification.

Fields

identifier: SerializedEnum

Expects an AssetIdentifier-compatible SerializedEnum.

asset_type: String

A name that must directly match one of the contract’s internal AssetDefinitionV3 names. Any request with a specified type not matching an asset definition will be rejected outright.

verifier_address: String

The bech32 address of a Verifier Account associated with the targeted AssetDefinitionV3, within its nested vector of VerifierDetailV2s.

access_routes: Option<Vec<AccessRoute>>

An optional parameter that allows the specification of a location to get the underlying asset data for the specified scope. The AccessRoute struct is very generic in its composition for the purpose of allowing various different solutions to fetching asset data. If the verification process requires generic lookups for each onboarded asset, access routes on the scope’s AssetScopeAttribute can be leveraged to easily determine the source of the underlying data. If these values are omitted at first, but later needed, they can always be added by using the UpdateAccessRoutes execution route. Note: Access routes can specify a name parameter, as well, to indicate the reason for the route, but this is entirely optional.

add_os_gateway_permission: Option<bool>

An optional parameter that will cause the emitted events to include values that signal to any Object Store Gateway watching the events that the selected verifier has permission to inspect the identified scope’s records via fetch routes. This will only cause a gateway to grant permissions to a scope to which the gateway itself already has read permissions. This essentially means that a key held by a gateway instance must have been used to store the scope’s records in Provenance Object Store.

This behavior defaults to TRUE.

VerifyAsset

This route is specifically designed to allow a Verifier specified in the AssetScopeAttribute of a Provenance Metadata Scope to indicate to the owner of the scope whether or not the content within the scope was valid or not. The Verifier Account, after determining validity of the underlying data, will either mark the classification as a success or failure. This route will reject all invokers except for Verifiers linked to a scope by the scope attribute, ensuring that only the verifier requested has the permission needed to classify an asset. In this way, the process for verification ensures that all involved parties’ requirements for security are satisfied. In addition, the verifier used in the process is stored on the scope attribute after the fact, ensuring that external inspectors of the generated attribute can choose which verifications to acknowledge and which to disregard.

Fields

identifier: SerializedEnum

Expects an AssetIdentifier-compatible SerializedEnum.

asset_type: String

The asset type this verification result is for

success: bool

A boolean indicating whether or not verification was successful. A value of false either indicates that the underlying data was fetched and it did not meet the requirements for a classified asset, or that a failure occurred during the verification process. Note: Verifiers should be wary of returning false immediately on a code failure, as this incurs additional cost to the onboarding account. Instead, it is recommended that verification implement some process that retries logic when exceptions or other code execution issues cause a failed verification.

message: Option<String>

An optional string describing the result of the verification process. If omitted, a standard message describing success or failure based on the value of success will be displayed in the AssetScopeAttribute.

access_routes: Option<Vec<AccessRoute>>

Like in the OnboardAsset message, this parameter allows the verifier to provide access routes for the assets that it has successfully fetched from the underlying scope data. This allows for the verifier to define its own subset of AccessRoute values to allow actors with permission to easily fetch asset data from a new location, potentially without any Provenance Blockchain interaction, facilitating the process of data interaction.

AddAssetDefinition

This route is only accessible to the contract’s admin address. This route allows a new AssetDefinitionV3 value to be added to the contract’s internal storage. These asset definitions dictate which asset types are allowed to be onboarded, as well as which verifiers are tied to each asset type. Each added asset definition must be unique in the following criteria:

Its asset_type value must not yet be registered in a different asset definition.

Fields

asset_definition: AssetDefinitionInputV3

An asset definition input value defining all of the new AssetDefinitionV3’s values. The execution route converts the incoming value to an asset definition.

UpdateAssetDefinition

This route is only accessible to the contract’s admin address. This route allows an existing AssetDefinitionV3 value to be updated. It works by matching the input’s asset_type to an existing asset definition and overwriting the existing values. If no asset definition exists for the given type, the request will be rejected.

Fields

asset_definition: AssetDefinitionInputV3

An asset definition input value defining all of the updated AssetDefinitionV3’s values. The execution route converts the incoming value to an asset definition.

ToggleAssetDefinition

This route is only accessible to the contract’s admin address. This route toggles an existing AssetDefinitionV3 from enabled to disabled, or disabled to enabled. When disabled, an asset definition will no longer allow new assets to be onboarded to the contract. Existing assets already onboarded to the contract and in pending status will still be allowed to be verified, but new values will be rejected. This same functionality could be achieved with an invocation of the UpdateAssetDefinition route but swapping the enabled value on the asset_definition parameter, but this route is significantly simpler and prevents accidental data mutation due to it not requiring the entirety of the definition’s values.

Fields

asset_type: String

The type of asset for which the definition’s enabled value will be toggled. As the asset type value on each asset definition is guaranteed to be unique, this key is all that is needed to find the target definition.

expected_result: bool

The value of enabled after the toggle takes place. This value is required to ensure that multiple toggles executed in succession (either by accident or by various unrelated callers) will only be honored if the asset definition is in the intended state during the execution of the route.

AddAssetVerifier

This route is only accessible to the contract’s admin address or the address of the verifier being updated. This route adds a new VerifierDetailV2 to an existing AssetDefinitionV3. This route is intended to register new verifiers without the bulky requirements of the UpdateAssetDefinition execution route. This route will reject verifiers added with addresses that match any other verifiers on the target asset definition.

Fields

asset_type: String

The type of asset for which the new VerifierDetailV2 will be added. This must refer to an existing AssetDefinitionV3’s asset_type value, or the request will be rejected.

verifier: VerifierDetailV2

The new verifier detail to be added to the asset definition, with all of its required values. No verifiers within the existing asset definition must have the same address value of this parameter, or the request will be rejected.

UpdateAssetVerifier

This route is only accessible to the contract’s admin address. This route updates an existing VerifierDetailV2 in an existing AssetDefinitionV3. This route is intended to be used when the values of a single verifier detail need to change, but not the entire asset definition. The request will be rejected if the referenced asset definition is not present within the contract, or if a verifier does not exist within the asset definition that matches the address of the provided verifier data.

Fields

asset_type: String

The type of asset for which the VerifierDetailV2 will be updated. This must refer to an existing AssetDefinitionV3’s asset_type value, or the request will be rejected.

verifier: VerifierDetailV2

The updated verifier detail to be modified in the asset definition. An existing verifier detail within the target asset definition must have a matching address value, or the request will be rejected.

UpdateAccessRoutes

This route is only accessible to the contract’s admin address OR to the owner of the access routes being updated. This route will swap all existing access routes for a specific owner for a specific scope to the provided values. These access routes either correspond to those created during the onboarding process, or those created during the verification process.

Fields

identifier: SerializedEnum

Expects an AssetIdentifier-compatible SerializedEnum.

asset_type: String

The asset type to update access routes for

owner_address: String

Corresponds to the bech32 address of the account that originally created the AccessRoutes. These values can be found in the AccessDefinition of the AssetScopeAttribute tied to a scope after the onboarding process occurs.

access_routes: Vec<AccessRoute>

A vector of AccessRoute to be used instead of the existing routes. If other existing routes need to be maintained and the updated is intended to simply add a new route, then the existing routes need to be included in the request alongside the new route(s).

DeleteAssetDefinition

This route is only accessible to the contract’s admin address. This route facilitates the removal of bad data. IMPORTANT: If an asset definition is completely removed, all contract references to it will fail to function. This can cause assets currently in the onboarding process for a deleted type to have failures when interactions occur with them. This functionality should only be used for an unused type!

Fields

asset_type: String

The asset type to delete the definition for