Function asset_classification_smart_contract::core::state::delete_fee_payment_detail

pub fn delete_fee_payment_detail<S1: Into<String>, S2: Into<String>>(
    storage: &mut dyn Storage,
    scope_address: S1,
    asset_type: S2,
) -> AssetResult<()>

Attempts to delete an existing payment detail by scope address. Returns an error if the detail does not exist or if deletion fails.

Parameters

storage A mutable reference to the contract’s internal storage.

scope_address The unique key scope_address of the detail to delete.