Function asset_classification_smart_contract::core::state::may_load_fee_payment_detail

pub fn may_load_fee_payment_detail<S1: Into<String>, S2: Into<String>>(
    storage: &dyn Storage,
    scope_address: S1,
    asset_type: S2,
) -> Option<FeePaymentDetail>

Attempts to find an existing fee payment detail by scope address, or returns a None variant if an error occurs or no detail is found.

Parameters

storage A reference to the contract’s internal storage.

scope_address The unique key scope_address for the requested payment detail.