Enum asset_classification_smart_contract::core::types::asset_identifier::AssetIdentifier

pub enum AssetIdentifier {
    AssetUuid(String),
    ScopeAddress(String),
}

An enum containing interchangeable values that can be used to define an asset (uuid or address).

Variants

AssetUuid(String)

A uuid v4 represented by a string.

ScopeAddress(String)

A bech32 Provenance Blockchain address that begins with “scope”.

Implementations

impl AssetIdentifier

pub fn from_serialized_enum(e: &SerializedEnum) -> AssetResult<Self>

Converts a SerializedEnum instance to one of the variants of this enum, if possible. On a failure, a ContractError::UnexpectedSerializedEnum error will be produced, indicating the unexpected value.

Parameters

e The serialized enum instance for which to attempt conversion.

pub fn to_serialized_enum(&self) -> SerializedEnum

Converts the specific variant of this enum to a SerializedEnum.

pub fn asset_uuid<S: Into<String>>(asset_uuid: S) -> Self

Creates a new instance of this enum as the AssetUuid variant.

Parameters

asset_uuid A uuid v4 string instance.

pub fn scope_address<S: Into<String>>(scope_address: S) -> Self

Creates a new instance of this enum as the ScopeAddress variant.

Parameters

scope_address A bech32 address with an hrp of “scope”.

pub fn get_asset_uuid(&self) -> AssetResult<String>

Fetches the asset uuid value from this enum. The AssetUuid variant can directly provide the value, but the ScopeAddress variant needs to utilize the scope_address_to_asset_uuid function to derive the value.

pub fn get_scope_address(&self) -> AssetResult<String>

Fetches the scope address value from this enum. The ScopeAddress variant can directly provide the value, but the AssetUuid variant needs to utilize the asset_uuid_to_scope_address function to derive the value.

pub fn to_identifiers(&self) -> AssetResult<AssetIdentifiers>

Takes the value provided and derives both values from it, where necessary, ensuring that both asset_uuid and scope_address are available to the user