Module asset_classification_smart_contract::core::types::serialized_enum

A simple struct that allows a type and value to be translated to some of the optional enums in the contract: AssetIdentifier, AssetQualifier, and ScopeSpecIdentifier.

Structs

SerializedEnum

There is a bug in cosmwasm 1.0.0’s interaction with serde-json-wasm that causes floating point operations to be added into the compiled wasm, so the previous solution of using things like AssetIdentifier directly and specifying them with a tag and content param in their serde annotation is impossible as of 1.0.0. This solution will allow existing requests to remain identical, but not generate floating point errors. It makes the schema less useful, but it’s a hack to fix a bug, so…