Function asset_classification_smart_contract::execute::add_asset_definition::add_asset_definition

pub fn add_asset_definition(
    deps: DepsMutC<'_>,
    env: Env,
    info: MessageInfo,
    msg: AddAssetDefinitionV1
) -> EntryPointResponse

The function used by execute when an ExecuteMsg::AddAssetDefinition message is provided. Attempts to add a new AssetDefinitionV3 to the contract’s internal storage.

Parameters

deps A dependencies object provided by the cosmwasm framework. Allows access to useful resources like contract internal storage and a querier to retrieve blockchain objects.

env An environment object provided by the cosmwasm framework. Describes the contract’s details, as well as blockchain information at the time of the transaction.

info A message information object provided by the cosmwasm framework. Describes the sender of the instantiation message, as well as the funds provided as an amount during the transaction.

msg An instance of the add asset definition v1 struct, provided by conversion from an ExecuteMsg.