Module asset_classification_smart_contract::execute::delete_asset_definition

Contains the functionality used by the DeleteAssetDefinition ExecuteMsg variant when invoked via the execute function.

Structs

DeleteAssetDefinitionV1

A transformation of ExecuteMsg::DeleteAssetDefinition for ease of use in the underlying delete_asset_definition function.

Functions

delete_asset_definition

Route implementation for ExecuteMsg::DeleteAssetDefinition. This function allows for the admin address to completely remove an asset definition. This is dangerous, because existing assets in the onboarding process for an asset definition will start emitting errors when being verified or retried. This should only ever be used on a definition that is guaranteed to be not in use and/or was erroneously added.