Function asset_classification_smart_contract::execute::delete_asset_definition::delete_asset_definition

pub fn delete_asset_definition(
    deps: DepsMutC<'_>,
    info: MessageInfo,
    msg: DeleteAssetDefinitionV1
) -> EntryPointResponse

Route implementation for ExecuteMsg::DeleteAssetDefinition. This function allows for the admin address to completely remove an asset definition. This is dangerous, because existing assets in the onboarding process for an asset definition will start emitting errors when being verified or retried. This should only ever be used on a definition that is guaranteed to be not in use and/or was erroneously added.

Parameters

deps A dependencies object provided by the cosmwasm framework. Allows access to useful resources like contract internal storage and a querier to retrieve blockchain objects.

info A message information object provided by the cosmwasm framework. Describes the sender of the instantiation message, as well as the funds provided as an amount during the transaction.

msg An instance of the delete asset definition v1 struct, provided by conversion from an ExecuteMsg.