Function asset_classification_smart_contract::execute::onboard_asset::onboard_asset

pub fn onboard_asset<'a, T>(
    repository: T, 
    info: MessageInfo, 
    msg: OnboardAssetV1
) -> EntryPointResponse where
    T: AssetMetaRepository + MessageGatheringService + DepsManager<'a>, 

The function used by execute when an ExecuteMsg::OnboardAsset message is provided. Attempts to verify that a provided Provenance Blockchain Metadata Scope is properly formed on a basic level, and then adds an AssetScopeAttribute to it as a Provenance Blockchain Attribute.

Parameters

repository A helper collection of traits that allows complex lookups of scope values and emits messages to construct the process of onboarding as a collection of messages to produce in the function’s result.

info A message information object provided by the cosmwasm framework. Describes the sender of the instantiation message, as well as the funds provided as an amount during the transaction.

msg An instance of the onboard asset v1 struct, provided by conversion from an ExecuteMsg.