Function asset_classification_smart_contract::execute::update_access_routes::update_access_routes

pub fn update_access_routes<'a, T>(
    env: &Env,
    repository: T,
    info: MessageInfo,
    msg: UpdateAccessRoutesV1,
) -> EntryPointResponse
where
    T: AssetMetaRepository + MessageGatheringService + DepsManager<'a>,

The function used by execute when an ExecuteMsg::UpdateAccessRoutes message is provided. Attempts to change the AccessRoutes for an AccessDefinition on a target AssetScopeAttribute.

Parameters

env An environment object provided by the cosmwasm framework. Describes the contract’s details, as well as blockchain information at the time of the transaction.

repository A helper collection of traits that allows complex lookups of scope values and emits messages to construct the process of updating access routes as a collection of messages to produce in the function’s result.

info A message information object provided by the cosmwasm framework. Describes the sender of the instantiation message, as well as the funds provided as an amount during the transaction.

msg An instance of the update access routes v1 struct, provided by conversion from an ExecuteMsg.