Function asset_classification_smart_contract::execute::update_asset_definition::update_asset_definition

pub fn update_asset_definition(
    deps: DepsMut<'_>,
    info: MessageInfo,
    msg: UpdateAssetDefinitionV1,
) -> EntryPointResponse

The function used by execute when an ExecuteMsg::UpdateAssetDefinition message is provided. Attempts to replace an existing AssetDefinitionV3 value based on a matching asset_type property. If no matching type is present, the request will be rejected.

Parameters

deps A dependencies object provided by the cosmwasm framework. Allows access to useful resources like contract internal storage and a querier to retrieve blockchain objects.

info A message information object provided by the cosmwasm framework. Describes the sender of the instantiation message, as well as the funds provided as an amount during the transaction.

msg An instance of the update asset definition v1 struct, provided by conversion from an ExecuteMsg.