Struct asset_classification_smart_contract::execute::update_asset_verifier::UpdateAssetVerifierV1

pub struct UpdateAssetVerifierV1 {
    pub asset_type: String,
    pub verifier: VerifierDetailV2,
}

A transformation of ExecuteMsg::UpdateAssetVerifier for ease of use in the underlying update_asset_verifier function.

Parameters

asset_type The unique identifier for the target AssetDefinitionV3, keyed on its asset_type property that the target verifier detail lives in, in its verifiers property.

verifier The verifier detail that will be updated. All values within this provided struct will replace the existing detail on the target AssetDefinitionV3.

Fields
asset_type: Stringverifier: VerifierDetailV2
Implementations

impl UpdateAssetVerifierV1

pub fn new<S: Into<String>>(asset_type: S, verifier: VerifierDetailV2) -> Self

Constructs a new instance of this struct.

Parameters

asset_type The unique identifier for the target AssetDefinitionV3, keyed on its asset_type property that the target verifier detail lives in, in its verifiers property.

verifier The verifier detail that will be updated. All values within this provided struct will replace the existing detail on the target AssetDefinitionV3.

pub fn from_execute_msg(msg: ExecuteMsg) -> AssetResult<UpdateAssetVerifierV1>

Attempts to create an instance of this struct from a provided execute msg. If the provided value is not of the UpdateAssetVerifier variant, then an InvalidMessageType error will be returned.

Parameters

msg An execute msg provided by the contract’s execute function.