Module asset_classification_smart_contract::execute::verify_asset

Contains the functionality used by the VerifyAsset ExecuteMsg variant when invoked via the execute function.

Structs

VerifyAssetV1

A transformation of ExecuteMsg::VerifyAsset for ease of use in the underlying verify_asset function.

Functions

verify_asset

The function used by execute when an ExecuteMsg::OnboardAsset message is provided. An execution route for use by the asset verifier selected by a requestor during the onboarding process to mark a scope as verified or rejected.