Module asset_classification_smart_contract::query

Contains the functionality used in the contract file to perform a contract query.

Modules

query_asset_definition

A query that fetches a target AssetDefinitionV3 from the contract’s internal storage.

query_asset_definitions

A query that fetches all AssetDefinitionV3s from the contract’s internal storage.

query_asset_scope_attribute

A query that attempts to find all AssetScopeAttributes on a Provenance Blockchain Metadata Scope that was added by this contract.

query_asset_scope_attribute_by_asset_type

A query that attempts to find an AssetScopeAttribute for a specific asset type on a Provenance Blockchain Metadata Scope that was added by this contract.

query_fee_payments

A query that attempts to find a FeePaymentDetail stored for an AssetScopeAttribute that has not yet finished its asset verification step.

query_state

A query that directly returns the contract’s stored StateV2 value.

query_version

A query that directly returns the contract’s stored VersionInfoV1 value.