Function asset_classification_smart_contract::query::query_asset_scope_attribute::query_scope_attribute_by_asset_uuid

pub fn query_scope_attribute_by_asset_uuid<S: Into<String>>(
    deps: &DepsC<'_>,
    asset_uuid: S,
) -> AssetResult<Vec<AssetScopeAttribute>>

Fetches an AssetScopeAttribute by the asset uuid value directly. Useful for internal contract functionality.

Parameters

deps A dependencies object provided by the cosmwasm framework. Allows access to useful resources like contract internal storage and a querier to retrieve blockchain objects.

asset_uuid Directly links to the asset_uuid value on an AssetScopeAttribute.