Function asset_classification_smart_contract::query::query_asset_scope_attribute_by_asset_type::may_query_scope_attribute_by_scope_address_and_asset_type

pub fn may_query_scope_attribute_by_scope_address_and_asset_type<S1: Into<String>, S2: Into<String>>(
    deps: &DepsC<'_>,
    scope_address: S1,
    asset_type: S2
) -> AssetResult<Option<AssetScopeAttribute>>

Fetches an AssetScopeAttribute by the scope address value directly for a particular asset type. The most efficient version of these functions, but still has to do a couple of lookups. This functionality should only be used on a once-per-transaction basis, if possible. Returns a ContractResult in the case of no attribute of the specified asset type being associated with the scope.

Parameters

deps A dependencies object provided by the cosmwasm framework. Allows access to useful resources like contract internal storage and a querier to retrieve blockchain objects.

scope_address Directly links to the scope_address value on an AssetScopeAttribute.

asset_type The asset type to query for