Function asset_classification_smart_contract::query::query_state::query_state

pub fn query_state(deps: &DepsC<'_>) -> AssetResult<Binary>

A query that directly returns the contract’s stored StateV2 value.

Parameters

deps A dependencies object provided by the cosmwasm framework. Allows access to useful resources like contract internal storage and a querier to retrieve blockchain objects.