Module asset_classification_smart_contract::service::asset_meta_repository

Defines a trait used for fetching and interacting with asset (Provenance Metadata Scope) values.

Traits

AssetMetaRepository

A trait used for fetching and interacting with asset (Provenance Metadata Scope) values.