Trait asset_classification_smart_contract::service::asset_meta_repository::AssetMetaRepository

pub trait AssetMetaRepository {
    fn has_asset<S1: Into<String>, S2: Into<String>>(
        &self,
        scope_address: S1,
        asset_type: S2
    ) -> AssetResult<bool>;
    fn onboard_asset(
        &self,
        env: &Env,
        attribute: &AssetScopeAttribute,
        verifier_detail: &VerifierDetailV2,
        is_retry: bool
    ) -> AssetResult<()>;
    fn update_attribute(
        &self,
        updated_attribute: &AssetScopeAttribute
    ) -> AssetResult<()>;
    fn get_asset<S1: Into<String>>(
        &self,
        scope_address: S1
    ) -> AssetResult<Vec<AssetScopeAttribute>>;
    fn try_get_asset<S1: Into<String>>(
        &self,
        scope_address: S1
    ) -> AssetResult<Option<Vec<AssetScopeAttribute>>>;
    fn get_asset_by_asset_type<S1: Into<String>, S2: Into<String>>(
        &self,
        scope_address: S1,
        asset_type: S2
    ) -> AssetResult<AssetScopeAttribute>;
    fn try_get_asset_by_asset_type<S1: Into<String>, S2: Into<String>>(
        &self,
        scope_address: S1,
        asset_type: S2
    ) -> AssetResult<Option<AssetScopeAttribute>>;
    fn verify_asset<S1: Into<String>, S2: Into<String>, S3: Into<String>>(
        &self,
        scope_address: S1,
        asset_type: S2,
        success: bool,
        verification_message: Option<S3>,
        access_routes: Vec<AccessRoute>
    ) -> AssetResult<AssetScopeAttribute>;
}

A trait used for fetching and interacting with asset (Provenance Metadata Scope) values.

Required Methods

fn has_asset<S1: Into<String>, S2: Into<String>>(
&self,
scope_address: S1,
asset_type: S2
) -> AssetResult<bool>

Determines if a scope exists with an AssetScopeAttribute emitted by this contract.

Parameters

scope_address The Provenance Metadata Scope bech32 address with an hrp of “scope” that refers to an existing scope.

asset_type The asset type to query for existence

fn onboard_asset(
&self,
env: &Env,
attribute: &AssetScopeAttribute,
verifier_detail: &VerifierDetailV2,
is_retry: bool
) -> AssetResult<()>

Attempts to generate the CosmosMsg values required to onboard an asset to the contract.

Parameters

attribute The scope attribute to be appended to the Provenance Metadata Scope as a result of a successful onboarding process.

verifier_detail The verifier chosen by the onboarding account.

is_retry Indicates that this onboarding action was attempted before, and the scope has an existing scope attribute with a failed verification on it.

fn update_attribute(
&self,
updated_attribute: &AssetScopeAttribute
) -> AssetResult<()>

Alters the internal values of the AssetScopeAttribute currently attached to a Provenance Metadata Scope with the provided values. This function will fail if no attribute exists on the target scope.

Parameters

updated_attribute The new attribute to attach to the scope. The original values will be entirely replaced with the values contained within this struct.

fn get_asset<S1: Into<String>>(
&self,
scope_address: S1
) -> AssetResult<Vec<AssetScopeAttribute>>

Attempts to fetch all asset attributes currently attached to a scope. Returns an error if no scope exists or no scope attribute is attached to the existing scope.

Parameters

scope_address A Provenance Blockchain bech32 address with an hrp of “scope”.

fn try_get_asset<S1: Into<String>>(
&self,
scope_address: S1
) -> AssetResult<Option<Vec<AssetScopeAttribute>>>

Attempts to fetch all asset attributes currently attached to a scope. Returns an error if bad data exists on the scope (an unlikely occurrence) or None if no scope or scope attributes exist.

Parameters

scope_address A Provenance Blockchain bech32 address with an hrp of “scope”.

fn get_asset_by_asset_type<S1: Into<String>, S2: Into<String>>(
&self,
scope_address: S1,
asset_type: S2
) -> AssetResult<AssetScopeAttribute>

Attempts to fetch an attribute currently attached to a scope by asset type. Returns an error if no scope exists or no scope attribute is attached to the existing scope.

Parameters

scope_address A Provenance Blockchain bech32 address with an hrp of “scope”.

asset_type The asset type to query for

fn try_get_asset_by_asset_type<S1: Into<String>, S2: Into<String>>(
&self,
scope_address: S1,
asset_type: S2
) -> AssetResult<Option<AssetScopeAttribute>>

Attempts to fetch asset attributes currently attached to a scope by asset type. Returns an error if bad data exists on the scope (an unlikely occurrence) or None if no scope or scope attribute exists.

Parameters

scope_address A Provenance Blockchain bech32 address with an hrp of “scope”.

asset_type The asset type to query for

fn verify_asset<S1: Into<String>, S2: Into<String>, S3: Into<String>>(
&self,
scope_address: S1,
asset_type: S2,
success: bool,
verification_message: Option<S3>,
access_routes: Vec<AccessRoute>
) -> AssetResult<AssetScopeAttribute>

Attempts to generate the CosmosMsg values required to verify an asset with the contract. Moves the AssetScopeAttribute to Approved

Parameters

scope_address A Provenance Blockchain bech32 address with an hrp of “scope”. Links to the desired scope to verify.

success Whether or not the scope should be considered verified when the process completes.

verification_message An optional value that will be displayed to external observers when fetching the AssetScopeAttribute associated with the scope.

access_routes Additional access routes that the verifier provides for external consumers to retrieve the underlying asset data from the scope, potentially without access an object store.