Trait asset_classification_smart_contract::service::message_gathering_service::MessageGatheringService

pub trait MessageGatheringService {
    // Required methods
    fn get_messages(&self) -> Vec<CosmosMsg>;
    fn add_message(&self, message: CosmosMsg);
    fn append_messages(&self, messages: &[CosmosMsg]);
    fn clear_messages(&self);
}

Specifies a trait used for dynamically aggregating CosmosMsg values without requiring the owning struct to be mutable.

Required Methods

fn get_messages(&self) -> Vec<CosmosMsg>

Retrieves all messages that have been appended to the service.

fn add_message(&self, message: CosmosMsg)

Moves an existing message into the service’s collection of messages.

fn append_messages(&self, messages: &[CosmosMsg])

Appends any number of existing messages by reference to the service.

fn clear_messages(&self)

Deletes all held messages from the service’s internal values.