Module asset_classification_smart_contract::util

Miscellaneous functionalities that do not logically belong to a concrete group.

Modules

aliases

Defines various types with type aliases to shorten syntax used elsewhere in the contract code.

constants

Defines all global constant values used throughout the contract.

contract_helpers

Functions that perform common actions for the execute, query, instantiate, and migrate functions.

deps_container

Allows dynamic delegation of a cosmwasm DepsMut to prevent common issues that arise when the struct is moved.

event_attributes

Helpers to ensure that emitting event attributes on execute calls occurs with standard values throughout the contract.

functions

Miscellaneous functions to use in various scenarios throughout the contract’s execution.

provenance_util

Utility functions that facilitate interaction with Provenance Blockchain modules.

scope_address_utils

Utility functions for interacting with bech32 addresses in the Provenance Blockchain environment.

traits

Global traits to be used across various areas of the contract.

vec_container

A container that allows a struct to manage a Vec, mutating its contents without it itself being declared as a mutable instance.