Enum asset_classification_smart_contract::util::event_attributes::EventType

pub enum EventType {
    InstantiateContract,
    MigrateContract,
    OnboardAsset,
    VerifyAsset,
    AddAssetDefinition,
    UpdateAssetDefinition,
    ToggleAssetDefinition,
    AddAssetVerifier,
    UpdateAssetVerifier,
    UpdateAccessRoutes,
    DeleteAssetDefinition,
}

An enum that contains all different event types that can occur throughout the contract’s routes.

Variants

InstantiateContract

Occurs when the contract is instantiated with instantiate.

MigrateContract

Occurs when the contract is migrated with migrate.

OnboardAsset

Occurs when the contract is executed to onboard an asset.

VerifyAsset

Occurs when the contract is executed to verify an asset.

AddAssetDefinition

Occurs when the contract is executed to add an asset definition.

UpdateAssetDefinition

Occurs when the contract is executed to update an asset definition.

ToggleAssetDefinition

Occurs when the contract is executed to toggle an asset definition.

AddAssetVerifier

Occurs when the contract is executed to add an asset verifier detail.

UpdateAssetVerifier

Occurs when the contract is executed to update an asset verifier detail.

UpdateAccessRoutes

Occurs when the contract is executed to update access routes.

DeleteAssetDefinition

Occurs when the contract is executed to delete an asset definition.

Implementations

impl EventType

pub fn event_name(self) -> String

Utilizes the implementation of Into to automatically derive the event name. This allows an invocation without an explicit type declaration.