Function asset_classification_smart_contract::util::functions::bank_send

pub fn bank_send<R: Into<String>, D: Into<String>>(
    recipient: R, 
    amount: u128, 
    denom: D
) -> CosmosMsg<ProvenanceMsg>

Creates a message that sends funds of the specified denomination from the contract to the recipient address. Important: The response type is of ProvenanceMsg, which allows this bank send message to match the type used for contract execution routes.

Parameters

recipient The bech32 address of the receiver of the sent funds.

amount An amount of coin to send (from the contract’s internal funding amount).

denom The denomination of coin to send.