Function asset_classification_smart_contract::util::functions::distinct_count_by_property

pub fn distinct_count_by_property<F, T, U>(slice: &[T], selector: F) -> usize where
    U: Sized + Eq + Hash,
    F: FnMut(&T) -> &U, 

Determines how many elements within the provided reference slice are unique by the given property.

Parameters

slice A reference slice from which to derive values to count.

selector A closure that defines the criteria used to determine when a value in the slice should be added to the count.

Examples

use asset_classification_smart_contract::util::functions::distinct_count_by_property;

let values = vec!["a", "b", "c", "a"];
let distinct_count = distinct_count_by_property(&values, |s| s);
assert_eq!(3, distinct_count);