Function asset_classification_smart_contract::util::functions::generate_asset_attribute_name

pub fn generate_asset_attribute_name<T: Into<String>, U: Into<String>>(
    asset_type: T,
    base_contract_name: U,
) -> String

Converts an asset type and a contract base name into an asset attribute that will be reserved to the contract for writing scope attributes.

Parameters

asset_type The value to use at the beginning of the name qualifier. Should refer to the asset_type property of an AssetDefinitionV3.

base_contract_name The value to use at the end of the name qualifier, after the dot. Should refer to the base_contract_name of the contract’s StateV2 internally-stored value.

Examples

use asset_classification_smart_contract::util::functions::generate_asset_attribute_name;

let asset_type = "mortgage";
let base_contract_name = "asset";
let attribute_name = generate_asset_attribute_name(asset_type, base_contract_name);
assert_eq!("mortgage.asset", attribute_name.as_str());