Function asset_classification_smart_contract::util::provenance_util::get_add_attribute_to_scope_msg

pub fn get_add_attribute_to_scope_msg(
    attribute: &AssetScopeAttribute, 
    base_contract_name: impl Into<String>
) -> AssetResult<CosmosMsg<ProvenanceMsg>>

Helper function to generate an “add attribute” message, as the functionality is re-used across multiple functions.

Parameters

attribute The scope attribute to be added to a Provenance Metadata Scope.

base_contract_name The base name of the contract, defined in the base_contract_name property of the StateV2 value stored internally in the contract.