Module asset_classification_smart_contract::util::scope_address_utils

Utility functions for interacting with bech32 addresses in the Provenance Blockchain environment.

Functions

asset_uuid_to_scope_address

Converts a string containing an asset uuid into a scope address.

bech32_string_to_addr

Validates that the address is valid by decoding to base 32, and then converts it to an Addr.

scope_address_to_asset_uuid

Takes a string representation of a scope address and converts it into an asset uuid string. Note: This conversion can also be called scope_address_to_scope_uuid because asset uuid always matches the scope uuid, as a convention.