Function asset_classification_smart_contract::util::scope_address_utils::scope_address_to_asset_uuid

pub fn scope_address_to_asset_uuid<S: Into<String>>(
    scope_address: S
) -> AssetResult<String>

Takes a string representation of a scope address and converts it into an asset uuid string. Note: This conversion can also be called scope_address_to_scope_uuid because asset uuid always matches the scope uuid, as a convention.

Parameters

scope_address A valid bech32 address with an hrp of “scope”.