Struct asset_classification_smart_contract::util::vec_container::VecContainer

pub struct VecContainer<T> {
    pub values: RefCell<Vec<T>>,
}

Contains a RefCell that manages a Vec of Clone-implementing values. Allows for the internal values to be mutated without declaring the value as mutable.

Fields
values: RefCell<Vec<T>>

The inner values of the container that are manipulated by the various implementation functions.

Implementations

impl<T> VecContainer<T>

pub fn new() -> Self

Construct a new instance of a container, starting with an empty vector.

pub fn push(&self, value: T)

Pushes a single owned item to the contained Vec.

Parameters

value A value append to the end of the inner Vec.

pub fn append(&self, values: &mut Vec<T>)

Appends an owned, mutable instance of a Vec containing instances of T to the contained Vec.

Parameters

values A mutable reference to an external vector of values to be added to the inner Vec.

pub fn clear(&self)

Removes all values from the inner Vec.

pub fn get(self) -> Vec<T>

Fetches the actual value inside the RefCell, moving the internalized value and disposing of this container in the process.

impl<T: Clone> VecContainer<T>

pub fn get_cloned(&self) -> Vec<T>

Fetches a cloned set of the owned values. Useful for early fetches without disposing of the container.

impl<T: Copy> VecContainer<T>

pub fn get_copied(&self) -> Vec<T>

Fetches a copied set of the owned values. Copying is more efficient than cloning, so this function should be preferred over get_cloned when an object contained in the VecContainer implements both.