Function asset_classification_smart_contract::validation::validate_execute_msg::validate_execute_msg

pub fn validate_execute_msg(msg: &ExecuteMsg) -> AssetResult<()>

The main branch of validation for an execute msg. Funnels the intercepted value based on variant to one of the various sub-functions in this module.

Parameters

msg An execute msg to process.