Module asset_classification_smart_contract::validation::validate_init_msg

Validates the integrity of an intercepted InitMsg and its associated AssetDefinitionV3 values.

Functions

validate_asset_definition

Validates that an asset definition value is properly formed, ensuring that all fields are properly set and fees are established correctly.

validate_asset_definition_input

Validates that an asset definition input value is properly formed, ensuring that all fields are properly set and fees are established correctly.

validate_init_msg

Validates the integrity of an intercepted InitMsg and its associated AssetDefinitionV3 values.

validate_verifier

Validates that a verifier detail is properly formed, ensuring that all fields are properly set and fees are established correctly.

validate_verifier_with_provided_errors

Validates that a verifier detail is properly formed, ensuring that all fields are properly set and fees are established correctly, with an additional funnel for exiting errors encountered beforehand.