Function asset_classification_smart_contract::validation::validate_init_msg::validate_asset_definition

pub fn validate_asset_definition(
    asset_definition: &AssetDefinitionV3
) -> AssetResult<()>

Validates that an asset definition value is properly formed, ensuring that all fields are properly set and fees are established correctly.

Parameters

asset_definition The asset definition value to validate for issues.