Enum asset_classification_smart_contract::core::error::ContractError

pub enum ContractError {

    Bech32Error(Error),
    SemVer(Error),
    Std(StdError),
    UuidError(Error),
    AssetAlreadyOnboarded {
        scope_address: String,
        asset_type: String,
    },
    AssetAlreadyVerified {
        scope_address: String,
        asset_type: String,
        status: AssetOnboardingStatus,
    },
    AssetNotFound {
        scope_address: String,
    },
    AssetPendingVerification {
        scope_address: String,
        asset_type: String,
        verifier_address: String,
    },
    AssetTypeDisabled {
        asset_type: String,
    },
    DuplicateVerifierProvided,
    GenericError {
        msg: String,
    },
    InvalidAddress {
        address: String,
        explanation: String,
    },
    InvalidContractName {
        current_contract: String,
        migration_contract: String,
    },
    InvalidContractVersion {
        current_version: String,
        migration_version: String,
    },
    InvalidFunds(String),
    InvalidMessageFields {
        message_type: String,
        invalid_fields: Vec<String>,
    },
    InvalidMessageType {
        expected_message_type: String,
    },
    InvalidScope {
        explanation: String,
    },
    NotFound {
        explanation: String,
    },
    RecordAlreadyExists {
        explanation: String,
    },
    RecordNotFound {
        explanation: String,
    },
    Unauthorized {
        explanation: String,
    },
    UnauthorizedAssetVerifier {
        scope_address: String,
        asset_type: String,
        verifier_address: String,
        expected_verifier_address: String,
    },
    UnexpectedSerializedEnum {
        received_type: String,
        explanation: String,
    },
    UnexpectedState {
        explanation: String,
    },
    Unimplemented,
    UnsupportedAssetType {
        asset_type: String,
    },
    UnsupportedVerifier {
        verifier_address: String,
        asset_type: String,
    },

}

A massive enum including all the different error scenarios that can be encountered throughout each process in the contract.

Variants

Bech32Error(Error)

An interceptor for a Bech32 Error.

SemVer(Error)

An interceptor for a SemVer Error.

Std(StdError)

An interceptor for a Cosmwasm Error.

UuidError(Error)

An interceptor for a Uuid Error.

AssetAlreadyOnboarded

This error is encountered when an asset is attempted in the onboarding process, but it has already been onboarded and classified.

Fields

scope_address: String

The bech32 scope address of the already-onboarded asset.

asset_type: String

The asset type for which onboarding has already been completed

AssetAlreadyVerified

An error emitted when a verifier attempts to run the verification process on an asset that does not require it. This can only occur when a verifier attempts to run a duplicate verification on a scope that has already had its verification process completed.

Fields

scope_address: String

The bech32 address of the scope that has already been verified.

asset_type: String

The asset type for which verification has already been completed for this asset.

status: AssetOnboardingStatus

The current onboarding status in the AssetScopeAttribute on the scope that has already been verified.

AssetNotFound

This error is encountered when the onboarding process cannot locate the scope specified by the requestor.

Fields

scope_address: String

The bech32 address of the scope that does not appear to yet exist on the Provenance Blockchain.

AssetPendingVerification

This error is encountered when the onboarding process is attempted for an asset that has already been onboarded as is awaiting a decision from its target verifier.

Fields

scope_address: String

The bech32 scope address of the asset that has been onboarded.

asset_type: String

The asset type for which verification is pending for this asset.

verifier_address: String

The bech32 address of the verifier that will perform verification on the asset.

AssetTypeDisabled

This error indicates that an asset was attempted to be onboarded with an asset_type linked to an AssetDefinitionV3 that is currently not enabled]

Fields

asset_type: String

The type of asset that is currently disabled.

DuplicateVerifierProvided

Denotes that an existing VerifierDetailV2 has the same address property as the provided VerifierDetailV2 to be added to an AssetDefinitionV3.

GenericError

An error that can be used in a circumstance where a named error is not necessary to be created.

Fields

msg: String

A free-form text description of the error that occurred.

InvalidAddress

Indicates that a bech32 address was provided that does not meet proper specifications for the given scenario.

Fields

address: String

The invalid bech32 address.

explanation: String

A message further explaining the issue.

InvalidContractName

An error that can occur during a migration that indicates that an incorrect stored contract code was attempted to be provided for a migration.

Fields

current_contract: String

The name of the existing contract. Should correlate to the value specified in Cargo.toml’s name property.

migration_contract: String

The name of the incorrect contract.

InvalidContractVersion

An error that can occur during a migration that indicates that the stored contract code that was used as the migration target has a version that is lower than the current contract version. This error is to prevent downgrading the contract to a previous version, which could remove features and cause problems.

Fields

current_version: String

The version of the contract currently active on the Provenance Blockchain.

migration_version: String

The version of the stored code used in the migration.

InvalidFunds(String)

A generic error that specifies that some form of provided or utilized coin was invalid.

Tuple Fields

0: String

Denotes the reason that invalid funds were detected.

InvalidMessageFields

An error emitted by the validation that indicates that input values were not properly specified (blank strings, invalid numbers, etc).

Fields

message_type: String

Indicates the type of message that was sent, causing the error. Will be one of the values in the msg file.

invalid_fields: Vec<String>

A collection of messages that indicate every issue present in the bad msg.

InvalidMessageType

An error emitted when an internal issue arises, indicating that an incorrect msg was sent to a function receiver. The functions that use this error are all at first called into in the contract file.

Fields

expected_message_type: String

Denotes the correct msg that was expected to be provided.

InvalidScope

An error that indicates that a scope inspected during the onboarding process is missing internal values and is not valid for onboarding, like an internal Provenance Blockchain Metadata Record.

Fields

explanation: String

A free-form text description of the reason that the scope is considered invalid.

NotFound

An error that occurs when a lookup is attempted for a contract resource but the resource does not exist. For instance, when an AssetDefinitionV3 does not contain a VerifierDetailV2 with a specified bech32 address, this error will occur.

Fields

explanation: String

A message describing the reason that the resource could not be found.

RecordAlreadyExists

An error that occurs when a unique key is violated during an attempt to add new data to the contract’s internal storage. Reference: state.

Fields

explanation: String

A free-form text description of the reason that the record already exists.

RecordNotFound

Occurs when a mandatory data lookup is performed on the contract’s internal storage, but the required value is not found. Reference: state.

Fields

explanation: String

A free-form text description of the record that could not be found.

Unauthorized

A generic error that occurs when an address attempts to perform an operation in the contract that it does not have the permission to.

Fields

explanation: String

A free-form text description of why the action was not authorized.

UnauthorizedAssetVerifier

An error emitted when an account attempts to initiate the verification process for a scope when its bech32 address is not listed as the verifier for the particular onboarding instance.

Fields

scope_address: String

The bech32 address of the scope that is awaiting verification.

asset_type: String

The asset type that was used to resolve the verifier

verifier_address: String

The bech32 address of the account that attempted to run verification.

expected_verifier_address: String

The bech32 address of the account that has been requested to run verification.

UnexpectedSerializedEnum

This error occurs when a SerializedEnum is received from a caller that cannot be properly converted to its expected underlying type.

Fields

received_type: String

The type value of the serialized enum that could not be properly converted.

explanation: String

A free-form text description of what went wrong in converting the value to its underlying type.

UnexpectedState

An error that can occur when the contract is in an unexpected state and refuses to perform an operation.

Fields

explanation: String

A free-form text description of why the operation was rejected.

Unimplemented

A placeholder error that can be used as a stopgap during the implementation of functionality to allow the tests to compile and run before the feature is completed. This error should never be used in a release build.

UnsupportedAssetType

This error is encountered when an asset type is attempted to be used during the onboarding process, but the asset type cannot be found in the contract’s internal storage.

Fields

asset_type: String

The type of asset that could not be located for onboarding.

UnsupportedVerifier

This error can occur when a target VerifierDetailV2 does not exist in an AssetDefinitionV3 during the onboarding process.

Fields

verifier_address: String

The bech32 address of the target verifier.

asset_type: String

The asset_type selected during onboarding.

Implementations

impl ContractError

pub fn generic<S: Into<String>>(msg: S) -> ContractError

Constructs an instance of the GenericError variant, allowing a generic Into implementation.

Parameters

msg The string value to use as the msg field of the generic error.