Function asset_classification_smart_contract::core::state::insert_fee_payment_detail

pub fn insert_fee_payment_detail(
    storage: &mut dyn Storage,
    fee_payment_detail: &FeePaymentDetail,
    asset_type: &str,
) -> AssetResult<()>

Inserts a new payment detail into storage. If a value already exists, an error will be returned. Note: Each payment detail must contain a unique scope_address value, or the insert will be rejected with an error.

Parameters

storage A mutable reference to the contract’s internal storage.

fee_payment_detail The detail to insert into storage and derive the unique scope address key.