Module asset_classification_smart_contract::core::types

Contains all structs used to drive core functionality throughout the contract.

Modules

access_definition

Defines a collection of AccessRoute for a specific address.

access_route

Defines a method of obtaining underlying asset data for a scope.

asset_definition

Defines a specific asset type associated with the contract. Allows its specified type to be onboarded and verified.

asset_identifier

An enum containing interchangeable values that can be used to define an asset (uuid or address).

asset_onboarding_status

An enum that denotes the various states that an AssetScopeAttribute can have.

asset_scope_attribute

An asset scope attribute contains all relevant information for asset classification, and is serialized directly as json into a Provenance Blockchain Attribute Module attribute on a Provenance Blockchain Metadata Scope.

asset_verification_result

A simple wrapper for the result of a verification for a scope.

entity_detail

Various fields describing an entity, which could be an organization, account, etc.

fee_destination

Defines an external account designated as a recipient of funds during the verification process.

fee_payment_detail

Defines a stored set of values for charging fees to the onboarding account during the asset classification process.

onboarding_cost

A node that defines how much onboarding should cost and any specific fees that should be paid.

serialized_enum

A simple struct that allows a type and value to be translated to some of the optional enums in the contract: AssetIdentifier

subsequent_classification_detail

Defines fees and values that can be used when classification is being done on an asset for a new type beyond the first.

verifier_detail

Defines the fees and addresses for a single verifier account for an AssetDefinitionV3.