Function asset_classification_smart_contract::execute::verify_asset::verify_asset

pub fn verify_asset<'a, T>(
    repository: T,
    info: MessageInfo,
    msg: VerifyAssetV1
) -> EntryPointResponsewhere
    T: AssetMetaRepository + MessageGatheringService + DepsManager<'a>,

The function used by execute when an ExecuteMsg::OnboardAsset message is provided. An execution route for use by the asset verifier selected by a requestor during the onboarding process to mark a scope as verified or rejected.

Parameters

repository A helper collection of traits that allows complex lookups of scope values and emits messages to construct the process of verification as a collection of messages to produce in the function’s result.

info A message information object provided by the cosmwasm framework. Describes the sender of the instantiation message, as well as the funds provided as an amount during the transaction.

msg An instance of the verify asset v1 struct, provided by conversion from an ExecuteMsg.