Module asset_classification_smart_contract::migrate::version_info

Module for structs and helper functions containing the version information stored for contract migrations.

Structs

VersionInfoV1

Holds both the contract’s unique name and version. Used to ensure that migrations have the correct targets and are not downgrades.

Constants

CONTRACT_NAME

Automatically derived from the Cargo.toml’s name property.

CONTRACT_VERSION

Automatically derived from the Cargo.toml’s version property.

Functions

get_version_info

Fetches, if possible, the current version information for the contract.

migrate_version_info

Sets the version info for the given contract to the derived values from the Cargo.toml file.

set_version_info

Sets the contract’s version definition directly to the specified VersionInfoV1 struct.