Struct asset_classification_smart_contract::service::asset_meta_service::AssetMetaService

pub struct AssetMetaService<'a> { /* private fields */ }

Ties all service code together into a cohesive struct to use for complex operations during the onboarding and verification processes.

Implementations

impl<'a> AssetMetaService<'a>

pub fn new(deps: DepsMutC<'a>) -> Self

Constructs a new instance of this struct.

Parameters

deps The cosmwasm deps that will be moved into a DepsContainer for future access.