Module asset_classification_smart_contract::util::constants

Defines all global constant values used throughout the contract.

Constants

ADDITIONAL_METADATA_KEY

Value = EventAdditionalMetadata meta string.

ASSET_EVENT_TYPE_KEY

Value = Event Type correlating to EvenType enum into String values (String).

ASSET_REGISTERED_KEY

Value = Asset UUID (String).

ASSET_SCOPE_ADDRESS_KEY

Value = Asset UUID (String).

ASSET_TYPE_KEY

Value = Asset Type (String).

NEW_ASSET_ONBOARDING_STATUS_KEY

Value = The new onboarding status of an AssetScopeAttribute after performing an execute function.

NEW_VALUE_KEY

Value = Any new value being changed that can be coerced to a string target. Dynamic to be used on various routes (String).

NHASH

A constant declaration to ensure the word “nhash” does not have typos when used throughout the contract’s source.

SCOPE_ID_KEY

Value = Scope ID Tied to the Asset (String).

SCOPE_OWNER_KEY

Value = The scope owner that sent the onboarding message.

VALID_VERIFIER_DENOMS

All denominations of coin that are valid for a verifier detail to include in its onboarding_denom field.

VERIFIER_ADDRESS_KEY

Value = The address of the verifier associated with the asset (String).