Function asset_classification_smart_contract::util::contract_helpers::check_admin_only

pub fn check_admin_only(deps: &Deps<'_>, info: &MessageInfo) -> AssetResult<()>

Ensures that only the admin of the contract can call into a route.

Parameters

deps A dependencies object provided by the cosmwasm framework. Allows access to useful resources like contract internal storage and a querier to retrieve blockchain objects.

info A message information object provided by the cosmwasm framework. Describes the sender of the instantiation message, as well as the funds provided as an amount during the transaction.

Examples

use cosmwasm_std::MessageInfo;
use asset_classification_smart_contract::util::contract_helpers::check_admin_only;
use cosmwasm_std::Addr;
use cosmwasm_std::testing::mock_info;
use provwasm_mocks::mock_provenance_dependencies;
use asset_classification_smart_contract::core::state::{STATE_V2, StateV2};

let mut deps = mock_provenance_dependencies();
STATE_V2.save(deps.as_mut().storage, &StateV2 { base_contract_name: "contract-name".to_string(), admin: Addr::unchecked("admin-name"), is_test: false })
    .expect("expected state to save successfully");
let info = mock_info("admin-name", &[]);
check_admin_only(&deps.as_ref(), &info).expect("admin-name was used as the admin and should return a success");