Struct asset_classification_smart_contract::util::event_attributes::EventAttributes

pub struct EventAttributes { /* private fields */ }

A helper struct to emit attributes for a Response.

Implementations

impl EventAttributes

pub fn new(event_type: EventType) -> Self

Constructs a new instance of this struct.

Parameters

event_type All events should denote their type for external consumers of Provenance Blockchain Event Stream, so this value is required for any new instance and appends the name of the event with the key of ASSET_EVENT_TYPE_KEY.

pub fn for_asset_event<T1: Into<String>, T2: Into<String>>( event_type: EventType, asset_type: T1, scope_address: T2, ) -> Self

Certain events like onboard_asset require a standard set of event types. This is a constructor for the struct that includes those values to facilitate the process of generating all events.

Parameters

event_type All events should denote their type for external consumers of Provenance Blockchain Event Stream, so this value is required for any new instance and appends the name of the event with the key of ASSET_EVENT_TYPE_KEY.

asset_type A unique key for an AssetDefinitionV3 that uses the key ASSET_TYPE_KEY.

scope_address A unique key for an AssetDefinitionV3 that uses the key ASSET_SCOPE_ADDRESS_KEY.

pub fn set_asset_type<T: Into<String>>(self, asset_type: T) -> Self

Appends an asset type value to an existing EventAttributes and returns the same instance to create a functional chain for further attribute addition.

Parameters

asset_type A unique key for an AssetDefinitionV3 that uses the key ASSET_TYPE_KEY.

pub fn set_scope_address<T: Into<String>>(self, scope_address: T) -> Self

Appends a scope address bech32 value to an existing EventAttributes and returns the same instance to create a functional chain for further attribute addition.

Parameters

scope_address A unique key for an AssetDefinitionV3 that uses the key ASSET_SCOPE_ADDRESS_KEY.

pub fn set_verifier<T: Into<String>>(self, verifier_address: T) -> Self

Appends a verifier address bech32 value to an existing EventAttributes and returns the same instance to create a functional chain for further attribute addition.

Parameters

verifier_address The address for a VerifierDetailV2 that uses the key VERIFIER_ADDRESS_KEY.

pub fn set_new_asset_onboarding_status( self, status: &AssetOnboardingStatus, ) -> Self

Appends an onboarding status value to an existing EventAttributes and returns the same instance to create a functional chain for further attribute addition.

Parameters

status The onboarding status of the current AssetScopeAttribute associated with the given event.

pub fn set_new_value<T: ToString>(self, new_value: T) -> Self

Appends a dynamic value to an existing EventAttributes and returns the same instance to create a functional chain for further attribute addition.

Parameters

new_value Any dynamic value that pertains to the current execution process, using the key NEW_VALUE_KEY.

pub fn set_scope_owner<T: ToString>(self, scope_owner: T) -> Self

Appends a scope owner bech32 value to an existing EventAttributes and returns the same instance to create a functional chain for further attribute addition.

Parameters

scope_owner A bech32 address that owns a Provenance Metadata Scope referred to by the current execution process, appended with the key SCOPE_OWNER_KEY.

pub fn set_additional_metadata( self, additional_metadata: &EventAdditionalMetadata, ) -> Self

Appends a dynamic set of additional metadata to an existing EventAttributes and returns the same instance to create a functional chain for further attribute addition. Note: If the metadata provided is empty, this key will be skipped to prevent strange value displays to external consumers.

Parameters

additional_metadata An instance of additional metadata to be displayed to any external consumers. Uses the key of ADDITIONAL_METADATA_KEY.