Module asset_classification_smart_contract::util::functions

Miscellaneous functions to use in various scenarios throughout the contract’s execution.

Functions

add_attribute

A helper to form a message for adding an attribute Adapted from provwasm-std.

add_json_attribute

A helper to form a message for adding a JSON attribute. Adapted from provwasm-std.

bank_send

Creates a message that sends funds of the specified denomination from the contract to the recipient address.

distinct_count_by_property

Determines how many elements within the provided reference slice are unique by the given property.

filter_valid_access_routes

Trims down a vector of AccessRoute to ensure that the contained values are valid and unique. Does the following: Ensures that access routes have a non-empty route property. Ensures that access routes either have an unset name, or a non-blank, set name. Ensures that all access routes, after being trimmed of trailing whitespace, are unique. Drops duplicates.

generate_asset_attribute_name

Converts an asset type and a contract base name into an asset attribute that will be reserved to the contract for writing scope attributes.

generate_os_gateway_grant_id

Converts an asset type and scope address into a grant id for use with Object Store Gateway. This combination will create a value unique to each verification’s process, ensuring that each selected verifier will always have access to its required scope values until any number of pending verifications complete.

msg_bind_name

Generates a name bind message that will properly assign the given name value to a target address. Assumes the parent name is unrestricted or that the contract has access to bind a name to the parent name. Adapted from funding-trading-bridge-smart-contract.

replace_single_matching_vec_element

Takes an existing vector, moves it into this function, swaps out a single existing item for a specified replacement item. If less or more than one existing item matches the given predicate closure, an error is returned.

try_into_add_attribute_request

Attempts to convert a [CosmosMsg] into a [MsgAddAttributeRequest]

try_into_bind_name_request

Attempts to convert a [CosmosMsg] into a [MsgBindNameRequest]

try_into_custom_fee_request

Attempts to convert a [CosmosMsg] into a [MsgAssessCustomMsgFeeRequest]

try_into_update_attribute_request

Attempts to convert a [CosmosMsg] into a [MsgUpdateAttributeRequest]

update_attribute

A helper to form a message for updating an existing attribute. Copied from provwasm-std.

validate_address

A helper that ensures address params are non-empty. Copied from provwasm-std.

validate_string

A helper that ensures string params are non-empty. Copied from provwasm-std.