Module asset_classification_smart_contract::util::functions

Miscellaneous functions to use in various scenarios throughout the contract’s execution.

Functions

bank_send

Creates a message that sends funds of the specified denomination from the contract to the recipient address. Important: The response type is of ProvenanceMsg, which allows this bank send message to match the type used for contract execution routes.

distinct_count_by_property

Determines how many elements within the provided reference slice are unique by the given property.

filter_valid_access_routes

Trims down a vector of AccessRoute to ensure that the contained values are valid and unique. Does the following: Ensures that access routes have a non-empty route property. Ensures that access routes either have an unset name, or a non-blank, set name. Ensures that all access routes, after being trimmed of trailing whitespace, are unique. Drops duplicates.

generate_asset_attribute_name

Converts an asset type and a contract base name into an asset attribute that will be reserved to the contract for writing scope attributes.

generate_os_gateway_grant_id

Converts an asset type and scope address into a grant id for use with Object Store Gateway. This combination will create a value unique to each verification’s process, ensuring that each selected verifier will always have access to its required scope values until any number of pending verifications complete.

replace_single_matching_vec_element

Takes an existing vector, moves it into this function, swaps out a single existing item for a specified replacement item. If less or more than one existing item matches the given predicate closure, an error is returned.