Function asset_classification_smart_contract::util::functions::generate_os_gateway_grant_id

pub fn generate_os_gateway_grant_id<T: Into<String>, U: Into<String>>(
    asset_type: T, 
    scope_address: U
) -> String

Converts an asset type and scope address into a grant id for use with Object Store Gateway. This combination will create a value unique to each verification’s process, ensuring that each selected verifier will always have access to its required scope values until any number of pending verifications complete.

Parameters

asset_type The value to use at the beginning of the grant id. Should refer to the asset_type property of an AssetDefinitionV3.

scope_address The bech32 address with a prefix of “scope” that uniquely defines the scope that is currently in the process of classification.